refactor: split sleep module by responsibility
- split batch.rs (3585→1752 lines) into:
  - prompt.rs: LLM response normalization/JSON extraction
  - extract.rs: Call1 event extraction
  - memory_update.rs: Call3 memory update (semantic+prospective)
  - rollup.rs: renamed from call2.rs, Call2 rollup planner+LLM
- rename prompt files:
  - sleep_batch_prompt_1.md → extract_prompt.md
  - sleep_batch_prompt_2.md → rollup_prompt.md
- batch.rs now orchestrates extract→rollup→memory_update
