+++ src/sleep/mod.rs
@@ -3,7 +3,10 @@
 pub(crate) mod scheduler;
 
 mod batch;
-mod call2;
 mod episodic_renderer;
+mod extract;
+mod memory_update;
+mod prompt;
+mod rollup;
 
 pub use batch::{SleepBatchError, run_sleep_batch};
